A method was implemented to protect the root directory for additions and deletions

+++ scripts/deldir.py
@@ -1,8 +1,11 @@
 import os
 import sys
+import json
 import shutil
+import hashlib
 import subprocess
 from pathlib import Path
+from getpass import getpass
 try:
     from rich.console import Console
 except ImportError as module:
@@ -16,6 +19,27 @@ guide_message = """With the deldir command, you can remove a directory in the cu
 
 Parameters:
 -f force to remove full directory"""
+
+
+# Get the root path.
+root_path = Path(__file__).parent.parent
+
+
+# Check root directory.
+def is_root(path):
+    return True if str(root_path) in str(path) else False
+
+
+# User authentication function.
+def authentication(password):
+    # Reading settings file.
+    with open(Path(root_path, "settings.json"), "r") as settings_file:
+        settings = json.load(settings_file)
+
+    sha_256 = hashlib.sha256()
+    sha_256.update(str(password).encode("UTF-8"))
+    hashed_password = sha_256.hexdigest()
+    return True if hashed_password == settings["password"] else False
 
 
 # Start-point.
@@ -39,9 +63,17 @@ def init():
             if dir_path.is_dir():
                 # Check if the directory is empty?
                 if len(os.listdir(dir_path)) <= 0:
-                    # If the directory is empty, it will be deleted.
-                    os.rmdir(dir_path)
-                    screen.print(f"'{dir_path.name}' directory successfully removed!", style="green")
+                    if is_root(dir_path):
+                        user_password = getpass("Enter password: ")
+                        if authentication(user_password):
+                            os.rmdir(dir_path)
+                            screen.print(f"'{dir_path.name}' directory successfully removed!", style="green")
+                        else:
+                            screen.print("Error: Authentication failed!", style="red")
+                    else:
+                        # If the directory is empty, it will be deleted.
+                        os.rmdir(dir_path)
+                        screen.print(f"'{dir_path.name}' directory successfully removed!", style="green")
                 else:
                     screen.print(f"Error: '{sys.argv[1]}' directory is full!", style="red")
             else:
@@ -59,9 +91,17 @@ def init():
         if dir_path.exists():
             # It is checked that the entry must be a directory.
             if dir_path.is_dir():
-                # Delete directory.
-                shutil.rmtree(dir_path)
-                screen.print(f"'{dir_path.name}' directory successfully removed!", style="green")
+                if is_root(dir_path):
+                    user_password = getpass("Enter password: ")
+                    if authentication(user_password):
+                        shutil.rmtree(dir_path)
+                        screen.print(f"'{dir_path.name}' directory successfully removed!", style="green")
+                    else:
+                        screen.print("Error: Authentication failed!", style="red")
+                else:
+                    # Delete directory.
+                    shutil.rmtree(dir_path)
+                    screen.print(f"'{dir_path.name}' directory successfully removed!", style="green")
             else:
                 screen.print("Error: You must choose a directory!", style="red")
         else:
