+++ scripts/delf.py
@@ -1,7 +1,10 @@
 import os
 import sys
+import json
+import hashlib
 import subprocess
 from pathlib import Path
+from getpass import getpass
 try:
     from rich.console import Console
 except ImportError as module:
@@ -9,6 +12,27 @@ except ImportError as module:
     from rich.console import Console
 finally:
     screen = Console()
+
+
+# Get the root path.
+root_path = Path(__file__).parent.parent
+
+
+# Check root directory.
+def is_root(path):
+    return True if str(root_path) in str(path) else False
+
+
+# User authentication function.
+def authentication(password):
+    # Reading settings file.
+    with open(Path(root_path, "settings.json"), "r") as settings_file:
+        settings = json.load(settings_file)
+
+    sha_256 = hashlib.sha256()
+    sha_256.update(str(password).encode("UTF-8"))
+    hashed_password = sha_256.hexdigest()
+    return True if hashed_password == settings["password"] else False
 
 
 # Start-point.
@@ -30,9 +54,17 @@ def init():
         if file_path.exists():
             # It is checked that the entry must be a file.
             if file_path.is_file():
-                # Delete file.
-                os.remove(file_path)
-                screen.print(f"'{file_path.name}' file successfully removed!", style="green")
+                if is_root(file_path):
+                    user_password = getpass("Enter password: ")
+                    if authentication(user_password):
+                        os.remove(file_path)
+                        screen.print(f"'{file_path.name}' file successfully removed!", style="green")
+                    else:
+                        screen.print("Error: Authentication failed!", style="red")
+                else:
+                    # Delete file.
+                    os.remove(file_path)
+                    screen.print(f"'{file_path.name}' file successfully removed!", style="green")
             else:
                 screen.print("Error: You must choose a file!", style="red")
         else:
